Move everything inside the Service
