+++ .jshintrc
@@ -2,7 +2,8 @@
   "predef": [
     "document",
     "window",
-    "-Promise"
+    "-Promise",
+    "Shepherd"
   ],
   "browser": true,
   "boss": true,
