--- app/components/ember-shepherd.js
@@ -1,2 +1,0 @@
-import EmberShepherd from 'ember-shepherd/components/ember-shepherd';
-export default EmberShepherd;
